+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0 2018-08-06
+
+* Support for old data format.
+* Data object is an instance of
+  `PolishVehicleRegistrationCertificateNewFormatData` or
+  `PolishVehicleRegistrationCertificateOldFormatData` class.
+
 ## v0.1.0 2018-07-28
 
 * Initial release